Remove unnecessary inheritance
