+++ firecrown/likelihood/source.py
@@ -536,57 +536,3 @@ class SourceGalaxy(Source, Generic[_SourceGalaxyArgsT]):
             z=z,
             dndz=nz,
         )
-
-
-@dataclass(frozen=True)
-class SourceCMBArgs:
-    """Class for CMB based sources arguments."""
-
-    scale: float = 1.0
-    field: str = "delta_matter"
-    z_source: float = 1100.0  # Add this line
-
-
-class SourceCMB(Source):
-    """Source class for CMB based sources."""
-
-    def __init__(
-        self,
-        *,
-        sacc_tracer: str,
-        scale: float = 1.0,
-    ):
-        """Initialize the SourceCMB object.
-
-        :param sacc_tracer: the name of the tracer in the SACC file.
-        :param scale: the scale of the source.
-        """
-        super().__init__(sacc_tracer)
-
-        self.sacc_tracer = sacc_tracer
-        self.scale = scale
-        self.current_tracer_args: None | SourceCMBArgs = None
-        self.tracer_args: SourceCMBArgs
-
-    def read_systematics(self, sacc_data: sacc.Sacc) -> None:
-        """Read the systematics for this source from the SACC file.
-
-        For CMB sources, there are no systematics to read.
-        """
-
-    def _read(self, sacc_data: sacc.Sacc) -> None:
-        """Read the CMB tracer data from a sacc file."""
-        if not hasattr(self, "tracer_args"):  # pragma: no cover
-            raise RuntimeError(
-                "Must initialize tracer_args before calling _read on SourceCMB"
-            )
-
-        # For CMB, we just verify the tracer exists
-        sacc_data.get_tracer(self.sacc_tracer)
-
-    def get_scale(self) -> float:
-        """Return the scale for this source."""
-        current_args = self.current_tracer_args
-        if current_args is None:
-            raise RuntimeError("current_tracer_args is not initialized")
-        return current_args.scale  # pylint: disable=no-member
